requirements.txt
locust
pyzmq==14.7.0
locustio==0.7.2
